[ALPS03639327] disp: fix coverity
[Detail]
fix coverity

MTK-Commit-Id: 2544d1b56353d996dffc4272c96b8d1969cc008d

CR-Id: ALPS03639327
Feature: others

+++ drivers/misc/mediatek/video/mt6771/dispsys/ddp_ovl.c
@@ -1303,7 +1303,7 @@ static int ovl_config_l(enum DISP_MODULE_ENUM module, struct disp_ddp_path_confi
 		}
 
 		Bpp = ufmt_get_Bpp(ovl_cfg->fmt);
-		tmp_bw = ovl_cfg->dst_h * ovl_cfg->dst_w * fps * Bpp;
+		tmp_bw = (unsigned long long)ovl_cfg->dst_h * ovl_cfg->dst_w * fps * Bpp;
 		do_div(tmp_bw, 1000);
 		tmp_bw *= 1250;
 		do_div(tmp_bw, fps * 1000 * 1000);
